regcomp.c: Generate new regnode for /[[:posix:]]/l
This follows on to the previous commit and generates code to use the new
regnode.  This allows some simplifications, with more to come in the
next commit.  The determination of the regnode is moved later in the
function that does this; the code that backed this out if we guessed
wrong is excised.